+++ src/main/java/com/troblecodings/invisiblelights/blocks/BlockLightBlocker.java
@@ -1,4 +1,4 @@
-package eu.gir.basics.blocks;
+package com.troblecodings.invisiblelights.blocks;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.BlockGetter;
